NLP-task/requirements.txt
matplotlib==3.2.1
pandas
requests
scikit-learn
tensorflow
tensorflow-addons
PyTorch
NumPy
matplotlib
